Change in Architecture
Instead of having a fire and forget type system, we do an _exchange_ of
messages. For every packet sent, we need to have a packet that was
recieved.

I thought this wouldn't be necessary, but reading from non-blocking
sockets doesn't work the way I thought it did. If a client sends
something over the network over to me, first up the selector might not
even select my socket for reading so I never know that there is
something available, and by the time I read, the packets can get
dropped, and the packets might take some time to transmit?
I honestly do not know. Here's a reproducer of what I'm
describing.

https://gist.github.com/ad-chaos/2f77f44441cfa7cbbdb405cd5125f6f2